auto_mail/requirements.txt
pandas
python-dotenv
openpyxl
numpy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib